pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "provide-workspace"
dynamic = ["version"]
description = "Development workspace setup for the provide.io ecosystem"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Provide", email = "support@provide.io" }
]

[project.urls]
Homepage = "https://foundry.provide.io/provide-workspace/"
Documentation = "https://foundry.provide.io/provide-workspace/"
Repository = "https://foundry.provide.io/provide-workspace/repo"
Issues = "https://foundry.provide.io/provide-workspace/issues"

# Optional workspace-level dependencies
# Individual projects manage their own dependencies
dependencies = [
    "provide-foundation[all]",
]

[dependency-groups]
dev = [
    "provide-testkit[all]",
]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.ruff]
line-length = 111
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
# Debug logging - enabled by default (ecosystem standard)
log_cli = true
log_cli_level = "DEBUG"

# Async configuration
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"

# Test discovery
testpaths = ["tests"]
pythonpath = ["src", "."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test execution
addopts = "-m 'not integration and not slow' -rFE -q --color=yes"

# Pytest markers
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring setup",
    "slow: Slow running tests",
    "fast: Fast tests taking <100ms",
]

# Warnings
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestAssertRewriteWarning",
]

# Directories to skip
norecursedirs = [
    ".git",
    ".hg",
    ".svn",
    "*_build",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    "venv",
    "workenv",
    "htmlcov",
    "docs/_build",
    ".hypothesis",
]

[tool.coverage.run]
branch = true
parallel = true
source = ["scripts"]

[tool.coverage.report]
show_missing = true
skip_covered = true
precision = 2
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "def __repr__",
    "def __str__",
    "pass",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
